Refactor: Enhance YouTube playlist synchronization and song removal
- Added `getSetVideoIdFromPlaylist` to Database for retrieving the set video ID associated with a song in a playlist.
- Improved `removeFromPlaylist` in YtMusic to accept `setVideoId` for more precise removal.
- Introduced `removeYTSongFromPlaylist` utility function to handle removal of songs from YouTube playlists, utilizing the set video ID.
- Modified `MediaItemMenu` to use the new `removeYTSongFromPlaylist` when removing songs from YouTube playlists, only delete song from local playlist if yt operation is successfully done.
- Modified `LocalPlaylistSongs` to use the new `removeYTSongFromPlaylist` when removing songs from YouTube playlists.
- Updated `Dialog` to use `removeYTSongFromPlaylist` to remove songs from YouTube playlists.
- Updated `PlayerMenu` to use `removeYTSongFromPlaylist` to remove songs from YouTube playlists.
- Updated `Utils` to save `setVideoId` when add to playlist.
- Updated `Utils` to delete song only if yt operation is successfully done.

+++ composeApp/src/androidMain/kotlin/it/fast4x/rimusic/ui/components/themed/Dialog.kt
@@ -165,6 +165,7 @@ import it.fast4x.rimusic.utils.isExplicit
 import it.fast4x.rimusic.utils.left
 import it.fast4x.rimusic.utils.lyricsSizeKey
 import it.fast4x.rimusic.utils.lyricsSizeLKey
+import it.fast4x.rimusic.utils.removeYTSongFromPlaylist
 import it.fast4x.rimusic.utils.right
 import it.fast4x.rimusic.utils.thumbnail
 import it.fast4x.rimusic.utils.thumbnailFadeExKey
@@ -2014,12 +2015,13 @@ fun SongMatchingDialog(
                                     .padding(vertical = 10.dp)
                                     .clickable(onClick = {
                                         Database.asyncTransaction {
-                                            deleteSongFromPlaylist(songToRematch.id, playlistId)
                                             if (isYouTubeSyncEnabled() && playlist?.isYoutubePlaylist == true && playlist.isEditable){
                                                 CoroutineScope(Dispatchers.IO).launch {
-                                                    YtMusic.removeFromPlaylist(playlist.browseId ?: "", songToRematch.id)
+                                                    if (removeYTSongFromPlaylist(songToRematch.id, playlist.browseId ?: "", playlistId))
+                                                        deleteSongFromPlaylist(songToRematch.id, playlistId)
                                                 }
                                             }
+
                                             if (songExist(song.asSong.id) == 0) {
                                                 Database.insert(song.asMediaItem)
                                             }
